groundtruthos-data/survey-automation/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "survey-automation"
version = "1.0.0"
description = "Survey automation pipeline for mixed survey inputs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "pyarrow>=16.0.0",
]

[project.scripts]
survey-automation = "survey_automation.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
